% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adapt.R
\name{adapt}
\alias{adapt}
\title{`adapt()` function - adapted for docker_moves}
\usage{
adapt(.runspec, .changes = NULL)
}
\arguments{
\item{.runspec}{path to runspec}

\item{.changes}{vector of file paths whose names match the table name they replace.}
}
\description{
This function adapts existing default data to a new custom input database.
As opposed to past versions, we just literally edit the default database, rather than creating a separate database.
This is more efficient, since the default database is already built in the docker image and replicable with every new container.
}
\author{
Tim Fraser
}
