% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{authorize}
\alias{authorize}
\title{Authenticate with Google Cloud using an API key json}
\usage{
authorize(key_path)
}
\arguments{
\item{key_path}{path to API key .json file}
}
\value{
A list object containing service account credentials.
}
\description{
Authenticate with Google Cloud using an API key json
}
