% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{bucket_delete}
\alias{bucket_delete}
\title{Delete Cloud Storage Bucket}
\usage{
bucket_delete(bucket, recursive = TRUE, token)
}
\arguments{
\item{bucket}{name of bucket for deletion (must be empty)}

\item{recursive}{TRUE}

\item{token}{...}
}
\description{
curl -X DELETE -H "Authorization: Bearer $(gcloud auth print-access-token)" \
}
