% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{bucket_list}
\alias{bucket_list}
\title{LIST BUCKETS}
\source{
https://cloud.google.com/storage/docs/listing-buckets#rest-list-buckets
}
\usage{
bucket_list(project = "projectname", token)
}
\arguments{
\item{project}{"moves_runs"}

\item{token}{...}
}
\description{
LIST BUCKETS
}
