% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{bucket_name}
\alias{bucket_name}
\title{Get name of bucket}
\usage{
bucket_name(dtablename, project = "projectname")
}
\arguments{
\item{project}{Name of project with linked billing ID - "projectname"}

\item{dbtablename}{name of eventual data table, eg. "d36109_u1_o12"}
}
\description{
Convert a dtablename into a bucket name
}
