% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{bucket_upload}
\alias{bucket_upload}
\title{Upload File to Bucket}
\usage{
bucket_upload(path, name, type, bucket, token)
}
\arguments{
\item{path}{Path of File to be uploaded}

\item{name}{Name for file to receive in bucket}

\item{type}{content-type, eg. "application/json" or "text/csv"}

\item{bucket}{eg. "projectname-d36109-u1-o12"}

\item{token}{...}
}
\description{
Upload File to Bucket
}
