% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{bucket_upload_bulk}
\alias{bucket_upload_bulk}
\title{Upload Objects in Bulk}
\usage{
bucket_upload_bulk(bucket, folder, last_file = NULL, token, scenario = FALSE)
}
\arguments{
\item{bucket}{...}

\item{token}{...}
}
\description{
Upload Objects in Bulk
}
