\name{cat_pav}
\alias{cat_pav}

\docType{data}

\title{Items parameters of the CAT_PAV data set (with item names)}

\description{
The CAT-PAV questionnaire (D.O. Santos, 2017) was developed to propose a computer-adaptive assessment of productive and contextualized academic English vocabulary, piloted with students at Iowa State University (USA). Items are presented as two sentences drawn from the academic subset of the COCA corpus, with one common missing word to be identified by test takers. In case of an incorrect first answer, synonyms are exhibited as a hint. Th item is scored 2 if a correct answer was provided from first attempt, 1 if a correct answer was provided after the hint being shown, and 0 in case of an incorrect answer after the hint was shown.
}

\format{
A matrix with 96 rows and three columns, respectively holding the discrimination, first threshold and second threshold under a GPCM calibration of the items of CAT-PAV study. Item names are available as row names, each item name corresponding to the word being tested by the item.
}

\source{
The original items of CAT-PAV were developed by D.O. Santos (2017), field tested at Iowa State University and calibrated using the Jmetrik software (Meyer, 2015).
}


\references{
Meyer, P. (2015). \emph{Jmetrik (version 4.0.3)} [Computer software]. Psychomeasurement Systems, LLC.

D.O. Santos, V. (2017). \emph{A computer-adaptive test of of productive and contextualized academic vocabulary breadth in English (CAT-PAV): Development and validation}. Graduate Theses and Dissertations, 16292. Ames, IA: Iowa State University. http://lib.dr.iastate.edu/etd/16292
}


\keyword{datasets}