% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{check_response}
\alias{check_response}
\title{Check and Report Response Status of a REST API Call}
\usage{
check_response(response)
}
\arguments{
\item{response}{REST API call response object}
}
\description{
Check and Report Response Status of a REST API Call
}
