% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{folder_to_files}
\alias{folder_to_files}
\title{Folder to Files}
\usage{
folder_to_files(folder, bucket, scenario = FALSE)
}
\arguments{
\item{folder}{name of folder, eg. "d36109_u1_o12"}
}
\description{
Convert folder name into a `tbl` of file paths, buckets, and content-types
}
\details{
#' @name bucket_create_old
#' @title Make Storage bucket
#' @note docker exec gcloud gcloud storage buckets create "gs://input-$BUCKET"
#' @param bucket_name ...
#' @param bucket_location https://cloud.google.com/storage/docs/locations
#' @param storage_class https://cloud.google.com/storage/docs/storage-classes
#' @param project should be "projectname"
#' @param token ...
#' @importFrom httr POST add_headers
#' @export
bucket_create = function(bucket_name, bucket_location = "us-central1", 
                         storage_class = 'STANDARD', project = "projectname", token) {
  
  # library(dplyr)
  # library(httr)
  # library(jsonlite)
  
  # Make json list data
  body = list(
    name = bucket_name,
    location = bucket_location,
    storageClass = storage_class,
    iamConfiguration = list(uniformBucketLevelAccess = list(enablled = TRUE))
  )
  
  # Prepare the request
  url = paste0("https://storage.googleapis.com/storage/v1/b?project=", project)
  headers = add_headers(
    "Authorization" = paste("Bearer ", token),
    "Content-Type" = "application/json"
  )
  # Send the POST request
  response <- POST(url = url, body = body, encode = "json", headers)
  
  # Check if request was successful
  check_response(response)
  
  
  # Return the response
  return(response)
}
}
