% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{format_data}
\alias{format_data}
\title{`format_data()`}
\usage{
format_data(
  .outputdbname = "moves",
  .run = NULL,
  .path = NULL,
  .level = "state",
  .geoid = "17",
  .by = c(16, 8, 12, 14, 15),
  .pollutant = NULL
)
}
\arguments{
\item{.outputdbname}{Name of output database. Assumes host is localhost.}

\item{.run}{# Specific MOVESRunID? (eg. 1)? If not, leave NULL.}

\item{.path}{character stringr showing path to save output. Defaults to NULL and self generates.}

\item{.level}{eg. "state"}

\item{.geoid}{eg. "17" (Illinois)}

\item{.by}{integer vector describing level of aggregation (eg. `8` = by sourcetype)}

\item{.pollutant}{Specific Pollutant (eg. 98)? If not, leave  NULL.}
}
\description{
Format MOVES output data into CAT data.
}
\author{
Tim Fraser
}
