% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bycats.R
\name{get_bycats}
\alias{get_bycats}
\title{Get `by` category variables}
\usage{
get_bycats(.by)
}
\arguments{
\item{.by}{integer id for aggregation level}
}
\description{
Returns a vector of columns corresponding to a given `by` value
}
