% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{get_new_bucket_name}
\alias{get_new_bucket_name}
\title{Get New Bucket Name}
\usage{
get_new_bucket_name(
  geoid = "36109",
  user = 1,
  project = "projectname",
  key_path = "runapikey.json",
  max_results = 100
)
}
\arguments{
\item{geoid}{eg. "36109"}

\item{user}{eg. 1}

\item{project}{eg. "projectname"}

\item{key_path}{eg. "runapikey.json"}

\item{max_results}{eg. 100}
}
\description{
We'll have trouble if we ask it to use a bucket name already in use. 
So, this function will return the next available bucket name for a given user.
}
