% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polprocesses.R
\name{get_particle}
\alias{get_particle}
\title{Get Particle-relevant cases of Required Pollutant-Process IDs.}
\usage{
get_particle(.pollutantID = 58)
}
\description{
Several species of primary exhaust PM2.5 all share the same prerequisites
This function will get the polProcessIDs for such cases more quickly.
}
\author{
Tim Fraser
}
