% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutantprocessassoc.R
\name{get_pollutantprocessassoc}
\alias{get_pollutantprocessassoc}
\title{Function to create pollutant process association chunks in runspec.}
\usage{
get_pollutantprocessassoc(.pollutants)
}
\arguments{
\item{.pollutants}{(integer vector) Vector of `pollutantID`s. Should be formatted as integers. (They will be coerced to be integers.)}
}
\description{
Vectorized function that returns a list of list objects
formatted to be pollutant-process association chunks in a Runspec.
}
\author{
Tim Fraser
}
