% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polprocesses.R
\name{get_polprocesses}
\alias{get_polprocesses}
\title{Get Required Pollutant-Process ID sets given an inputted `.pollutantID`.}
\usage{
get_polprocesses(.pollutantID = 98)
}
\arguments{
\item{.pollutantID}{(integer) pollutant ID}
}
\description{
For any `.pollutantID` supplied, this function will return a character `vector` of 
all Pollutant-Process IDs required to measure that pollutant's emissions.
This corresponds to clicking on a Pollutant
in the "Selected" column of the MOVES GUI, 
then clicking "Select Prerequisites".

NOTE: This does NOT allow you to select a specific pollutant-process 
THEN get all required derivative pollutant-processes.
This only works with an overall pollutant as an input, where it assumes 
you want to estimate that pollutant for ALL its relevant pollutant processes.
}
\author{
Tim Fraser
}
