% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polprocesses.R
\name{get_polProcessID}
\alias{get_polProcessID}
\title{Get Pollutant-Process ID}
\usage{
get_polProcessID(.pollutantID, .processID)
}
\description{
Concatenate pollutant and process ID information
Given a supplied pollutant ID and .processID, give me all required polProcessIDs
}
\note{
Functions for getting pollutant information
}
\author{
Tim Fraser
}
