% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutantprocessassoc.R
\name{get_pp_attributes}
\alias{get_pp_attributes}
\title{Get Pollutant-Process Attributes from a Runspec chunk, provided as a `R` `list()` object}
\usage{
get_pp_attributes(.list)
}
\arguments{
\item{.list}{list object, eg. produced by `set_pp_attributes()`}
}
\description{
Convert to table from a list formatted as a pollutant-process chunk from a Runspec
}
\author{
Tim Fraser
}
