% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roadtypes.R
\name{get_roadtypes}
\alias{get_roadtypes}
\title{Get `<roadtypes>` Runspec chunk}
\usage{
get_roadtypes(.roadtypes = NULL)
}
\arguments{
\item{.roadtypes}{(integer) Vector of integer `roadTypeID`s. Defaults to `NULL`, which selects all `roadTypeID`s. Recommended to use `NULL` unless you have a compelling reason.}
}
\description{
Get `<roadtypes>` Runspec chunk
}
\examples{
get_roadtypes(.roadtypes = c(1, 2))

}
\author{
Tim Fraser
}
