% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_runid.R
\name{get_runid}
\alias{get_runid}
\title{`get_runid()` function}
\usage{
get_runid(n = 1, id = TRUE, outputdbname = "moves")
}
\arguments{
\item{n}{number of most recent runs to return data for. Default is 1. Max is 100.}

\item{id}{TRUE}

\item{outputdbname}{"moves"}
}
\description{
Function to get the MOVESRunID for your most recent moves run.
}
