% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polprocesses.R
\name{get_simple}
\alias{get_simple}
\title{Get Simple Cases of Required Pollutant-Process IDs}
\usage{
get_simple(.pollutantID = 67)
}
\description{
Many Metals, Dioxins, and Furans are just running exhaust (XX1)
plus Total Gaseous Hydrocarbons (101)
This function will get the polProcessIDs for such cases more quickly.
}
\author{
Tim Fraser
}
