% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vehicleselections.R
\name{get_v_attributes}
\alias{get_v_attributes}
\title{Get Onroad Vehicle Selection attributes from a `R` `list()` as a table.}
\usage{
get_v_attributes(.list)
}
\description{
Get Onroad Vehicle Selection attributes from a `R` `list()` as a table.
}
\author{
Tim Fraser
}
