% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{job_get}
\alias{job_get}
\title{Get Job Details}
\usage{
job_get(job_name, project = "projectname", region = "us-central1", token)
}
\arguments{
\item{job_name}{name of job}

\item{project}{"projectname"}

\item{region}{"us-central1"}

\item{token}{...}
}
\description{
curl -H "Content-Type: application/json" \
-H "Authorization: Bearer ACCESS_TOKEN" \
-X GET \
-d '' \
https://run.googleapis.com/v2/projects/PROJECT_ID/locations/REGION/jobs/JOB-NAME
}
