% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{object_delete}
\alias{object_delete}
\title{Delete Object from Storage Bucket}
\usage{
object_delete(file, bucket, token)
}
\arguments{
\item{file}{file for deletion from bucket}

\item{bucket}{name of bucket eg. "projectname-d36109-u1-o12"}

\item{token}{...}
}
\description{
Delete Object from Storage Bucket
}
