% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{object_delete_bulk}
\alias{object_delete_bulk}
\title{Delete Objects in Bulk}
\usage{
object_delete_bulk(bucket, token)
}
\arguments{
\item{bucket}{...}

\item{token}{...}
}
\description{
Delete Objects in Bulk
}
