% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{object_list}
\alias{object_list}
\title{List Objects in a Google Cloud Storage Bucket}
\usage{
object_list(bucket, token)
}
\arguments{
\item{bucket}{name of storage bucket, eg. 'projectname-d36109-u1-o12'}

\item{token}{...}
}
\description{
List Objects in a Google Cloud Storage Bucket
}
