% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutantprocessassoc.R
\name{pp_to_pollutant}
\alias{pp_to_pollutant}
\title{Pollutant-Process ID to Pollutant ID}
\usage{
pp_to_pollutant(.polProcessID)
}
\arguments{
\item{`.polProcessID`}{(integer/character) pollutant process id eg. 601}
}
\description{
Convert a `polProcessID` to just the `pollutantID`
}
\author{
Tim Fraser
}
