% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutantprocessassoc.R
\name{pp_to_process}
\alias{pp_to_process}
\title{Pollutant-Process ID to Process ID}
\usage{
pp_to_process(.polProcessID)
}
\arguments{
\item{`.polProcessID`}{(integer/character) pollutant process id eg. 601}
}
\description{
Convert a `polProcessID` to just the `processID`
}
\author{
Tim Fraser
}
