% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_activity.R
\name{process_activity}
\alias{process_activity}
\title{`process_activity()`}
\usage{
process_activity(tab, .by, .geoid)
}
\arguments{
\item{tab}{...}

\item{.by}{...}

\item{.geoid}{...}
}
\description{
Function to process activity data for a specific .geoid and .by, 
given a 'tab' - a local connection to a specific table, 
which is obtained formatted in the process_by.R function.
Meant to be used as a sub-function.
}
\author{
Tim Fraser, March 2023
}
