% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_by.R
\name{process_by}
\alias{process_by}
\title{`process_by()`}
\usage{
process_by(
  con,
  .by = 16,
  .geoid = "36109",
  .level = "county",
  .run = NULL,
  .pollutant = NULL
)
}
\arguments{
\item{con}{...}

\item{.by}{...}

\item{.geoid}{...}

\item{.level}{...}

\item{.run}{...}

\item{.pollutant}{...}
}
\description{
Meta-function to process emissions and activity data, by invoking the process_emissions.R and process_activity.R scripts.
Process data from a local MariaDB connection 'con' for one '.by' group, for a specified .geoid at a given .level
Can specify a specific .pollutant if you only want one pollutant. Otherwise, .pollutant = NULL runs all pollutants by default
Can specific a specific .run if there are multiple runs in your database. Otherwise, just takes the contents of your database
}
\author{
Tim Fraser, March 2023
}
