% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_emissions.R
\name{process_emissions}
\alias{process_emissions}
\title{`process_emissions()`}
\usage{
process_emissions(tab, .by, .geoid, .pollutant = NULL)
}
\arguments{
\item{tab}{...}

\item{.by}{...}

\item{.geoid}{...}

\item{.pollutant}{...}
}
\description{
This function allows the user/session to connect to one of our frequently used database connections, depending on input criteria.
Function to process emissions data for a specific .geoid and .by and .pollutant, (though by default, .pollutant is NULL and we gather data for all pollutants), given a 'tab' - a local connection to a specific table, which is obtained formatted in the process_by.R function.
Meant to be used as a sub-function.
}
\author{
Tim Fraser, March 2023
}
