% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{read_api}
\alias{read_api}
\title{Read contents of a REST API call response}
\usage{
read_api(response)
}
\arguments{
\item{response}{REST API call response object}
}
\description{
Read contents of a REST API call response
}
