% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_to_parameters.R
\name{rs_to_parameters}
\alias{rs_to_parameters}
\title{Make parameters from runspec}
\usage{
rs_to_parameters(
  path_rs = "rs_custom.xml",
  path_parameters = "parameters.json",
  tablename = NULL,
  user = 1,
  order = 1,
  by = c(1, 16, 8, 12, 14, 15),
  multiple = FALSE,
  return = FALSE
)
}
\arguments{
\item{path_rs}{(character) path to runspec .xml file. Should be named `rs_custom.xml`, at whatever path needed.}

\item{path_parameters}{(character) path to parameters .json file. Should be named `parameters.json`, at whatever path needed.}

\item{tablename}{(character) eg. "d36109_u1_o1" name of Cloud SQL table to make.
Alternatively, you can provide the `user` id and `order` id instead and it will make a `tablename` for you.}

\item{user}{(integer) userid, eg. `1`}

\item{order}{(integer) orderid, eg. `2`}

\item{by}{(integer vector) aggregation type ids. Ranges from `1` to `16`. Defaults are `1`,`16`,`8`,`12`,`14`, and `15`.}

\item{return}{(logical) Return `p`, the parameters list? Default is `FALSE`.}
}
\description{
Make parameters from runspec
}
\author{
Tim Fraser
}
