% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pollutantprocessassoc.R
\name{set_pp_attributes}
\alias{set_pp_attributes}
\title{Set Pollutant-Process Attributes in a Runspec as an `R` `list()` object}
\usage{
set_pp_attributes(.pollutantID, .pollutantName, .processID, .processName)
}
\arguments{
\item{.pollutantID}{eg. 98}

\item{.pollutantName}{eg. CO2 Equivalent}

\item{.processID}{eg. 1}

\item{.processName}{eg. Running Exhaust}
}
\description{
Creates 1 list item with 4 attributes.
This will become 1 pollutant-process chunk in a Runspec file.
}
\author{
Tim Fraser
}
