% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_roadtypes.R
\name{set_r_attributes}
\alias{set_r_attributes}
\title{Set Roadtype Attributes for a `<roadtype>` Runspec chunk.}
\usage{
set_r_attributes(.roadtypeid, .roadtypename, .modelcombination)
}
\description{
Set Roadtype Attributes for a `<roadtype>` Runspec chunk.
}
