% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vehicleselections.R
\name{set_v_attributes}
\alias{set_v_attributes}
\title{Set Onroad Vehicle Selection attributes in `R` `list()` format.}
\usage{
set_v_attributes(.fueltypeid, .fueltypedesc, .sourcetypeid, .sourcetypename)
}
\description{
Set Onroad Vehicle Selection attributes in `R` `list()` format.
}
\author{
Tim Fraser
}
