\name{tcals}
\alias{tcals}

\docType{data}

\title{Items parameters of the TCALS 1998 data set}

\description{
 The TCALS (\emph{Test d'Aptitude en Anglais Langue Seconde}) is an aptitude test of English language as a second language
 in the French speaking college of Outaouais (Gatineau, QC, Canada). The test consists of 85 items and is administered 
 every year to newly incoming students. The item parameters of the year 1998 have been estimated under the 3PL model.
 Inattention parameters are therefore fixed to one.
}

\format{
 A matrix with 85 rows and four columns, respectively holding the discrimination, difficulty, pseudo-guessing a,nd inattention parameters 
 as calibrated on the results of the 1998 application of the TCALS questionnaire.
 }

\source{
 The TCALS test was originally developed by Laurier, Froio, Pearo and Fournier (1998) and item parameters were obtained from Raiche (2002)..
}


\references{
 Laurier, M., Froio, L., Pearo C., and Fournier, M. (1998). Test de classement d'anglais langue seconde au collegial. 
 Montreal, Canada: College de Maisonneuve.

 Raiche, G. (2002). Le depistage du sous-classement aux tests de classement en anglais, langue seconde, au collegial
 [The detection of under classification to the collegial English as a second language placement tests].
 Gatineau, QC: College de l'Outaouais.
}


\keyword{datasets}