% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_functions.R
\name{workflow_execute}
\alias{workflow_execute}
\title{Execute Workflow}
\usage{
workflow_execute(
  workflow_name,
  body,
  project = "moves-runs",
  region = "us-central1",
  token
)
}
\description{
Primarily used to publish tables from `granddata` to `orderdata`, for Cornell CATCLOUD developers.
`body` should be a list, matching this structure:
`list(fromdbname = "orderdata", fromdbtable = "d36109_u1_o18", todbname = "granddata", todbtable =  "dZZZZZ")`
}
\author{
Tim Fraser
}
